Tests: allow keepeing logs

+++ test/test-event-upd3.sh
@@ -74,6 +74,6 @@ echo $res | grep "$year-$month-01 11:00:00"
 echo $res | grep "$year-$month-01 11:15:00"
 echo $res | grep "$resource"
 
-rm -f $test_log
+if [-n "$delete_log"]; then rm -f $test_log; fi
 
 exit 0
