+++ test/test-recur-2rdate.sh
@@ -76,6 +76,6 @@ echo $res | grep "$year-$month-13 03:30:00"
 echo $res | grep "$year-$month-13 03:45:00"
 echo $res | grep "$resource"
 
-rm -f $test_log
+if [-n "$delete_log"]; then rm -f $test_log; fi
 
 exit 0
