+++ test/test-recur-cancel-all.sh
@@ -72,6 +72,6 @@ res=`$psql -c \
 (echo $res | grep "$uid") && exit 3
 (echo $res | grep "$resource") && exit 4
 
-rm -f $test_log
+if [-n "$delete_log"]; then rm -f $test_log; fi
 
 exit 0
